requirements.txt
apache-libcloud==2.8.1
appnope==0.1.0
apptools==4.5.0
APScheduler==3.6.3
asgiref==3.2.5
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
bokeh==2.0.0
bs4==0.0.1
cdsapi==0.2.7
certifi==2019.11.28
cftime==1.1.1.2
chardet==3.0.4
cloudpickle==1.3.0
configobj==5.0.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-extensions==2.2.8
docutils==0.16
entrypoints==0.3
envisage==4.9.2
future==0.18.2
futures==3.1.1
gitdb==4.0.5
GitPython==3.1.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
ipyevents==0.7.1
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
jellyfish==0.6.1
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
keyring==21.2.1
kiwisolver==1.2.0
llvmlite==0.31.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
packaging==20.3
pandas==1.0.2
pandas-market-calendars==1.2
pandocfilters==1.4.2
parso==0.6.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pip==20.1.1
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2==2.8.5
ptyprocess==0.6.0
pycryptodome==3.9.7
pyface==6.1.2
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.6.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
readme-renderer==26.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scipy==1.4.1
seaborn==0.10.0
see19==0.3.5
selenium==3.141.0
Send2Trash==1.5.0
setuptools==47.1.1
sh==1.13.1
six==1.14.0
smmap==3.0.4
sortedcontainers==2.1.0
soupsieve==2.0
sqlparse==0.3.1
statsmodels==0.11.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
traits==6.0.0
traitsui==6.1.3
tweepy==3.8.0
twine==3.1.1
typing-extensions==3.7.4.1
tzlocal==2.1
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.8
us==2.0.2
vtk==8.1.2
wcwidth==0.1.8
webencodings==0.5.1
wheel==0.34.2
widgetsnbextension==3.5.1
xarray==0.15.1
xlrd==1.2.0
zipp==3.1.0
